requirements.txt
requests-html
aioredis
pyyaml
click
Jinja2
aiohttp
sender
